Move Direction and associates to kas::data module
